\name{cor3}
\alias{cor3}
\title{Calc: correlation after excluding NA/Inf/Blank}
\usage{
cor3()
}
\description{
Calculating correlation coefficient betwen X and Y
}
\examples{

library(aj412s); library(readxl)

Adata<-read_excel('자료1.xlsx')

cor2(data$x, data$y)
cor2(data$x, data$y)
cor2(data$x, data$y)
cor2(data$x, data$y)


}