\name{desc2}
\alias{desc2}
\title{Calc. descriptive statistics}
\usage{
desc2()
}
\description{
Calculating descriptive statistics.
}
\examples{
> desc2(Adata$가격만족)
               기술통계결과
자료수               16.000
NA갯수                1.000
NA제거후자료수       15.000
중위수                4.000
평균                  3.733
표준편차              0.961
분산                  0.924
최대값                5.000
최소값                2.000
변동계수SM            0.257
Q1사분위                3
Q2사분위                4
Q3사분위                4

}