\name{s_rim}
\alias{s_rim}
\title{RIM valuation}
\usage{
s_rim()
}
\description{
Stock Valuation
}
\examples{
# 입력 자료 출처: finance.naver.com -> 종목분석 -> 기업현황
# s_rim('21,613', 12.51, 8, '46,957,120')
# 예시) 데이터 입력 예시: s_rim( '21,613', 12.51, 8, '46,957, 120) 
# B0=지배주주지분(단위 억원)= 2.1613e+12  ROE= 0.1251  Ke= 0.08  Nstock=주식수= 46957120 
# 적정주가 =  71974.88 
# 초과이익 10% 감소 시 적정주가 =  56406.21 
# 초과이익 20% 감소 시 적정주가 =  51958.02 

}
